third_party/blink/web_tests/platform/linux-chrome/external/wpt/fetch/stale-while-revalidate/stale-css-expected.txt

This is a testharness.js-based test.
[FAIL] Cache returns stale resource
  assert_equals: (revalidated version loaded) expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
